メッセージ取得メソッドを分離
