+++ resources/ts/sendMessage.ts
@@ -6,29 +6,6 @@ const btnSendMessage = <HTMLButtonElement>
 const inputMessage = <HTMLInputElement>
 	document.querySelector('#input-message');
 
-const getMessageList = () => {
-	axios.get('/messages/get')
-		.then((result) => {
-			const resultObj = result.data;
-			console.log(resultObj);
-			for (let index = 0; index < resultObj.length; index++) {
-				const message = resultObj[index];
-				initMessageList(message.body);
-			}
-		})
-		.catch((error) => {
-			console.log('error!');
-			console.log(error.data);
-		});
-};
-
-const initMessageList = (message: string) => {
-	let listBlock = document.querySelector('#list-block');
-	let list: HTMLLIElement = document.createElement('li');
-	list.innerText = message;
-	listBlock?.appendChild(list);
-}
-
 const addMessageList = (addMessage: string): void => {
 	let listBlock = document.querySelector('#list-block');
 	let list: HTMLLIElement = document.createElement('li');
@@ -58,6 +35,4 @@ const listenMessageChannel = () => {
 }
 
 btnSendMessage.addEventListener("click", sendMessage);
-
-document.addEventListener('DOMContentLoaded', getMessageList);
 document.addEventListener('DOMContentLoaded', listenMessageChannel);
